+++ src/main/java/com/example/miniprojectbe/entity/BasketId.java
@@ -1,4 +1,4 @@
-package com.example.miniprojectbe.dto;
+package com.example.miniprojectbe.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
@@ -6,12 +6,11 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
-
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 @EqualsAndHashCode
-public class PurchaseId implements Serializable {
+public class BasketId implements Serializable {
 
     private String member;
     private Long item;
