+++ src/main/java/com/example/miniprojectbe/entity/Purchase.java
@@ -1,6 +1,5 @@
 package com.example.miniprojectbe.entity;
 
-import com.example.miniprojectbe.dto.PurchaseId;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
