+++ src/main/java/com/example/miniprojectbe/entity/PurchaseId.java
@@ -1,4 +1,4 @@
-package com.example.miniprojectbe.dto;
+package com.example.miniprojectbe.entity;
 
 import lombok.AllArgsConstructor;
 import lombok.EqualsAndHashCode;
@@ -6,11 +6,12 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 
 import java.io.Serializable;
+
 @NoArgsConstructor
 @AllArgsConstructor
 @Getter
 @EqualsAndHashCode
-public class BasketId implements Serializable {
+public class PurchaseId implements Serializable {
 
     private String member;
     private Long item;
